Turn on HAVE_BUILTIN_TYPES_COMPATIBLE_P and fix typing mistakes
C's implicit promotion rules make min/max somewhat dangerous if
the types are not matched, fix all the errors.

This should be squished into the patch introducing ccan

+++ libibverbs/examples/rc_pingpong.c
@@ -580,8 +580,8 @@ static int pp_post_send(struct pingpong_context *ctx)
 }
 
 struct ts_params {
-	unsigned int		 comp_recv_max_time_delta;
-	unsigned int		 comp_recv_min_time_delta;
+	uint64_t		 comp_recv_max_time_delta;
+	uint64_t		 comp_recv_min_time_delta;
 	uint64_t		 comp_recv_total_time_delta;
 	uint64_t		 comp_recv_prev_time;
 	int			 last_comp_with_ts;
